+++ lib/or_tools/sum_array.rb
@@ -17,7 +17,7 @@ module ORTools
     end
 
     def to_s
-      "(#{@array.map(&:to_s).join(" + ")})"
+      "#{@array.map(&:to_s).reject { |v| v == "0" }.join(" + ")}"
     end
   end
 end